Rename test and wrap the block theme shortcuts
